feat(broot): add some binds
broot is really cool. more familiar with yazi, but i see where i *want*
broot. image preview for PDFs might be nice, but yazi excels in that
department. for yazi, i want to extend the typical 'z' (for fzf)
binding, and write a previewer-command; that should mimic some of the
qualities i like about broot, especially if i can truncate results from
fd. it'd be interesting to just add broot as a file selector, for deeply
nested files. smart-filter is already a great step in the right
direction, but its inability to go backwards (with some keybinding) is
not acceptable; might write my own plugin